clean up backend structure
